requirements.txt
flappy-bird-gym==0.3.0
neat-python==0.92